% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_offset.R
\name{apply_offset}
\alias{apply_offset}
\title{EITrans::apply_offset}
\usage{
apply_offset(ens, offset, pre_sorted = F, verbose = F)
}
\arguments{
\item{ens}{A 4-dimensional array for ensemble forecasts. Dimensions should be
\verb{[stations, times, lead times, members]}.}

\item{offset}{Member offset values used to calibrate ensembles.}

\item{pre_sorted}{Whether the ensemble members are pre-sorted from the lowest to
the highest. Using pre-sorted ensembles can save runtime.}

\item{verbose}{Whether to print messages.}
}
\value{
A calibrated ensemble
}
\description{
EITrans::apply_offset adds the offset values to a forecast ensemble. Offset values
are calculated from EITrans::member_offset. Or you could simply use EITrans::EITrans
which has built in a complete workflow for EITrans calibration.
}
