% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalInverse.R
\name{empiricalInverse}
\alias{empiricalInverse}
\title{EITrans::empiricalInverse}
\usage{
empiricalInverse(AnEn, config, final.ensemble.size, LOO.test.size,
  LOO.ensemble.size, historical.similar.count = 1, keep.table = T,
  keep.LOO.rank = T, keep.LOO = T, do.not.append = F,
  show.progress = T, silent = F, i.station = NA, i.test.day = NA,
  AnEn.LOO.test = NA, member.name = "analogs")
}
\arguments{
\item{AnEn}{The AnEn results from \code{\link{generateAnalogs}}.}

\item{config}{The configuration used by \code{\link{generateAnalogs}}.}

\item{final.ensemble.size}{How many ensemble members to keep in the
final ensemble forecasts.}

\item{LOO.test.size}{The number of historical forecasts to
carry out AnEn leave-one-out test.}

\item{LOO.ensemble.size}{The number of ensemble members to take from
the LOO test results to construct the reference rank histogram.}

\item{historical.similar.count}{The number of past similar forecasts
to include when considering calculating the reference rank histogram.}

\item{keep.table}{Whether to keep the summary table of similariy.}

\item{keep.LOO.rank}{Whether to keep the verification rank histogram
for the leave-one-out tests.}

\item{keep.LOO}{Whether to keep the analogs and similarity results
from the leave-one-out test if they are computed.}

\item{do.not.append}{If set to True, results will not be appended to
AnEn, and the filtered analog ensembles will be returned. At the
same time, all \code{keep*} parameters become invalid.}

\item{show.progress}{Whether to show the progress bar.}

\item{silent}{Whether to be silent.}

\item{i.station}{A vector of station indices of which analogs will be filtered.}

\item{i.test.day}{A vector of test day indices of which analogs will be filtered.}

\item{AnEn.LOO.test}{The precomputed AnEn results for leave-one-out tests.
If not provided, LOO tests will be carried out at the spot. See details for
instructions on precopmuting LOO tests.}

\item{member.name}{The name of the member in AnEn to process. By default, it
is \code{analogs}. But this parameter can be helpful if, for example, you want to
process another member called \code{analogs.alternative}.}
}
\description{
EITrans::empiricalInverse is designed to improve the quality of
Analog Ensemble forecasts by using empirical inverse tranformation
function (EITrans). EITrans works better with ensembles
with large amount of members (usually at the level of a hunderd).
}
\details{
To precompute AnEn.LOO.test, simply copy the values from \code{search_times_compare}
to \code{test_times_compare} in the \code{config} which is passed to this function, and
use the \code{generateAnalogs} to generate the \code{AnEn.LOO.test}. To save memory and space,
you can optionally omit similarity information in the results by setting
\code{config$keep_similarity = F}. \strong{The only required member is \code{analogs}}.

You might wonder when you need to generate an updated version of \code{AnEn.LOO.test}.
Usually, if you change the settings in \code{config} which would affect the search
data, for example, \code{search_times_compare}, and the calculation of \code{AnEn}, for
example, \code{weights}, you will need to regenerate the \code{AnEn.LOO.test}. Otherwise, you
do not need to regenerate it. For example, you don't need to regenerate \code{AnEn.LOO.test}
if you only change \code{test_times_compare}.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
