% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/index_to_value.R
\name{index_to_value}
\alias{index_to_value}
\title{EITrans::index_to_value}
\usage{
index_to_value(similarity_time_index, ens, obs = NULL, verbose = F)
}
\description{
EITrans::index_to_value
}
\author{
Weiming Hu \href{mailto:weiming@psu.edu}{weiming@psu.edu}
}
