% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_offset.R
\name{member_offset}
\alias{member_offset}
\title{EITrans::EITrans}
\usage{
member_offset(
  ensembles,
  observations,
  left_delta,
  right_delta,
  infinity_estimator = 1,
  verbose = T,
  pre_sorted = F
)
}
\arguments{
\item{ensembles}{A 4-dimensional array with the dimensions
\verb{[stations, times, forecast lead times, ensemble members]}}

\item{observations}{A 3-dimensional array with the dimensions
\verb{[stations, times, forecast lead times]}}

\item{left_delta}{A scalar addition offset for the minimum during sampling}

\item{right_delta}{A scalar addition offset for the maximum during sampling}

\item{infinity_estimator}{A scalar to scaling the spread of the ensemble}

\item{verbose}{Whether to be verbose}

\item{pre_sorted}{Whether the input ensemble members are pre-sorted. Use pre-sorted
ensembles will save you about 80\% of the execution time. To pre-sort your ensembles,
use \code{aperm(apply(ensembles, 1:3, sort, na.last = T), c(2, 3, 4, 1))}.}
}
\value{
A vector for ensemble member offset. \strong{Make sure that you have sorted
your ensemble values before you add the offset to each member}.
}
\description{
EITrans::member_offset calculates an value offset for each member of the input
ensembles so that the resulting rank histogram shall be flat.
}
\details{
Input ensembles and observations should match up for the first three dimensions
which should be #stations, #times, and #lead times. The 4th dimension of ensembles
should be ensemble members.

The length of the output of the function will be a vector of the length of the
ensemble members. When you actually use these offset values to correct new ensembles,
\strong{make sure that you SORT your ensemble members from smallest to largest}.
}
\author{
Weiming Hu \href{mailto:weiming@psu.edu}{weiming@psu.edu}
}
