% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_parameters.R
\name{optimize_parameters}
\alias{optimize_parameters}
\title{EITrans::optimize_parameters}
\usage{
optimize_parameters(grid_search, ens_train, obs_train, verbose = F)
}
\arguments{
\item{grid_search}{A data frame with columns \code{left_deltas}, \code{right_deltas},
\code{left_infinity}, \code{right_infinity}, and \code{multiplier}.}

\item{ens_train}{A 4-dimensional array as training ensemble forecasts.
Dimensions should be \verb{[stations, times, lead times, members]}.}

\item{obs_train}{A 3-dimensional array for observations that correspond to the
ensemble forecasts. Dimensions should be \verb{[stations, times, lead times]}.}

\item{verbose}{Whether to print progress information.}
}
\value{
A list with member offsets calculated with each of the parameter combination
and the rank histogram of the calibrated training forecast ensemble with the
corresponding parameter combination.
}
\description{
EITrans::optimize_parameters carries out a search for the optimized parameter
combination for EITrans calibration with the input training ensembles forecasts
and observations.
}
\author{
Weiming Hu \href{mailto:weiming@psu.edu}{weiming@psu.edu}
}
