% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_ensemble_analogs.R
\name{univariate_ensemble_analogs}
\alias{univariate_ensemble_analogs}
\title{EITrans::univariate_ensemble_analogs}
\usage{
univariate_ensemble_analogs(
  ens,
  ens_times,
  ens_flts,
  ens_times_train,
  ens_times_dev,
  circular_ens = F,
  config = NULL
)
}
\arguments{
\item{ens}{A 4-dimensional array for ensemble forecasts. Dimensions should be
\verb{[stations, times, lead times, members]}.}

\item{ens_times}{A vector for ensemble forecast times.}

\item{ens_flts}{A vector for ensemble forecast lead times.}

\item{ens_times_train}{Training times from the ensemble forecast times.}

\item{ens_times_dev}{Development times from the ensemble forecast times.}

\item{circular_ens}{Whether the ensemble forecast variable is circular.}
}
\description{
EITrans::univariate_ensemble_analogs generates analogs using univariate ensemble
forecasts. It only returns the time indices where the most similar historical
ensemble forecasts are found.
}
\author{
Weiming Hu \href{mailto:weiming@psu.edu}{weiming@psu.edu}
}
