% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boostEnsembleSize.R
\name{boostEnsembleSize}
\alias{boostEnsembleSize}
\title{EITrans::boostEnsembleSize}
\usage{
boostEnsembleSize(AnEn, config, AnEn.LOO.test = NA, scale.size = 10,
  show.progress = T, silent = F, avoid.duplicates = T)
}
\description{
EITrans::boostEnsembleSize boosts the number of ensemble memebers using
an empirical approach.
}
\details{
An important assumption is that AnEn.LOO.test is generated using the same config
as AnEn except for the config$test_times_compare and config$num_members being changed.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
