% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EITrans.R
\name{EITrans}
\alias{EITrans}
\title{EITrans::EITrans}
\usage{
EITrans(
  ens,
  ens_times,
  ens_flts,
  ens_times_train,
  ens_times_dev,
  ens_times_test,
  obs,
  left_deltas,
  right_deltas,
  infinity_estimator,
  multiplier,
  circular_ens = F,
  member_weights = NULL,
  grid_search_cores = detectCores()
)
}
\arguments{
\item{ens}{A 4-dimensional array for ensemble forecasts. Dimensions should be
\verb{[stations, times, lead times, members]}.}

\item{ens_times}{A vector for ensemble forecast times.}

\item{ens_flts}{A vector for ensemble forecast lead times.}

\item{ens_times_train}{Training times from the ensemble forecast times.}

\item{ens_times_dev}{Development times from the ensemble forecast times.}

\item{ens_times_test}{Testing times from the ensemble forecast times.}

\item{obs}{A 3-dimensional array for observations that correspond to the
ensemble forecasts. Dimensions should be \verb{[stations, times, lead times]}.}

\item{left_deltas}{A vector of left-edge deltas to experiment.}

\item{infinity_estimator}{A vector of values to experiment for estimating
the ensemble spread.}

\item{multiplier}{A vector of values to experiment for adjusting the
ensemble member offset.}

\item{circular_ens}{Whether the ensemble forecast variable is circular.}

\item{member_weights}{Weights for each ensemble members when finding similar
historical ensemble forecasts.}

\item{grid_search_cores}{The number of cores to use during grid search.}

\item{rigth_deltas}{A vector of right-edge deltas to experiment.}
}
\value{
A list with the calibrated ensemble forecasts and intermediate results.
}
\description{
EITrans::EITrans is the main function for ensemble forecast calibration.
}
\author{
Weiming Hu \href{mailto:weiming@psu.edu}{weiming@psu.edu}
}
