% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inverseTransform.R
\name{inverseTransform}
\alias{inverseTransform}
\title{EITrans::inverseTransform}
\usage{
inverseTransform(heuristic.rank, analogs, members.to.keep, digits = 6,
  randomize.selection = getOption("randomize.selection", T))
}
\arguments{
\item{heuristic.rank}{The frequency or count for each ranked bin.
An example of this parameter is the list member \emph{rank} generated
using the function \code{\link{verifyRankHist}}.}

\item{analogs}{A 4-dimensional array for analog values. An exmple
of this parameter is the list member \emph{analogs} with only the \emph{value}
column generated using the function \code{\link{generateAnalogs}}.}

\item{members.to.keep}{The number of ensemble members to keep.}

\item{digits}{The number of decimals when computing the cumulous summation.}

\item{randomize.selection}{Since rank bins are determined by member values,
a decision needs to be made whether to choose the left edge or right edge value
as the cut value when a specific bin is chosen. Set this to True to
enable randomized selection of left or right edges. Set this to False to
always select the left edge. If this function is called by other functions,
you can set the parameter by \code{options(randomize.selection = F)}.}
}
\description{
EITrans::inverseTransform resample the analog ensemble members
based on the distribution match information from the heuristic rank
histogram. This function keeps a subset of members from the original
forecast ensembles and the goal is to improve the reliability of
the ensemble subset.
}
\author{
Weiming Hu \email{weiming@psu.edu}
}
